+++ Constraid/ExpandFromEdge.swift
@@ -5,8 +5,7 @@
 #endif
 
 /**
- Constrains the object's leading edge to be greater than or equal to the
- leading edge of `item`
+ Constrains the object's leading edge to expand outward from the leading edge of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
  - parameter itemB: The `item` you want to constrain itemA against
@@ -28,15 +27,15 @@ public func expand(_ itemA: Constraid.View, fromLeadingEdgeOf itemB: Any?,
     itemA.translatesAutoresizingMaskIntoConstraints = false
     let collection = Constraid.ConstraintCollection([
         NSLayoutConstraint(item: itemA, attribute: .leading,
-                           relatedBy: .greaterThanOrEqual, toItem: itemB,
+                           relatedBy: .lessThanOrEqual, toItem: itemB,
                            attribute: .leading, multiplier: multiplier,
-                           constant: offset, priority: priority)
+                           constant: (offset * -1), priority: priority)
         ])
     return collection
 }
 
 /**
- Constrains the object's trailing edge to be greater than or equal to the
+ Constrains the object's trailing edge to expand outward from the
  leading edge of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
@@ -62,13 +61,13 @@ public func expand(_ itemA: Constraid.View, fromTrailingEdgeOf itemB: Any?,
         NSLayoutConstraint(item: itemA, attribute: .trailing,
                            relatedBy: .greaterThanOrEqual, toItem: itemB,
                            attribute: .trailing, multiplier: multiplier,
-                           constant: (offset * -1), priority: priority)
+                           constant: offset, priority: priority)
         ])
     return collection
 }
 
 /**
- Constrains the object's top edge to be greater than or equal to the
+ Constrains the object's top edge to outward from the
  top edge of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
@@ -91,14 +90,14 @@ public func expand(_ itemA: Constraid.View, fromTopEdgeOf itemB: Any?,
     itemA.translatesAutoresizingMaskIntoConstraints = false
     let collection = Constraid.ConstraintCollection([
         NSLayoutConstraint(item: itemA, attribute: .top,
-                           relatedBy: .greaterThanOrEqual, toItem: itemB, attribute: .top,
-                           multiplier: multiplier, constant: offset, priority: priority)
+                           relatedBy: .lessThanOrEqual, toItem: itemB, attribute: .top,
+                           multiplier: multiplier, constant: (offset * -1), priority: priority)
         ])
     return collection
 }
 
 /**
- Constrains the object's bottom edge to be greater than or equal to the
+ Constrains the object's bottom edge to outward from the
  bottom edge of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
@@ -124,14 +123,14 @@ public func expand(_ itemA: Constraid.View, fromBottomEdgeOf itemB: Any?,
         NSLayoutConstraint(item: itemA, attribute: .bottom,
                            relatedBy: .greaterThanOrEqual, toItem: itemB,
                            attribute: .bottom, multiplier: multiplier,
-                           constant: (offset * -1), priority: priority)
+                           constant: offset, priority: priority)
         ])
     return collection
 }
 
 /**
- Constrains the object's top & bottom edges to be greater than or equal
- to the top & bottom edge of `item`
+ Constrains the object's top & bottom edges to expand outward from the
+ top & bottom edges of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
  - parameter itemB: The `item` you want to constrain itemA against
@@ -159,8 +158,8 @@ public func expand(_ itemA: Constraid.View, fromHorizontalEdgesOf itemB: Any?,
 }
 
 /**
- Constrains the object's leading & trailing edges to be greater than or equal
- to the leading & trailing edge of `item`
+ Constrains the object's leading & trailing edges to expand outward from
+ the leading & trailing edge of `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object
  - parameter itemB: The `item` you want to constrain itemA against
@@ -187,8 +186,8 @@ public func expand(_ itemA: Constraid.View, fromVerticalEdgesOf itemB: Any?,
 }
 
 /**
- Constrains the object's top, bottom, leading & trailing edges to be
- greater than or equal to the top, bottom, leading & trailing edge of
+ Constrains the object's top, bottom, leading & trailing edges to expand
+ outward from the top, bottom, leading & trailing edge of
  `item`
 
  - parameter itemA: The `item` you want to constrain in relation to another object